Initial Commit - Form Validation PHP
Form Validation  upon submission. 

Satisfies the following conditions:
Create a PHP script that parses an HTML form with the following user data and constraints.
Email (valid email address)
Username (6-10 characters; alphanumeric only)
Password (at least one lowercase letter, one uppercase letter, and one number)
Gender (either “male,” “female,” "non-binary," or “other”)
Use a PHP object as the source for this data (in lieu of a database)
The validation should be server-side and handled on form submission (client-side validation is out of scope for this assignment).
The form should be valid HTML5 that meets the W3C standard.
If the user enters bad input, your script should display an error message and repopulate the form so that the user can change their entered data.
Note that the Password field should be exempt from repopulation. It should always be blank on every form render.
Upon successful validation, write the data to a log file.